Fix FPS limiter: stop treating the refresh rate as a frame cap
Selecting a refresh rate fed the FPS limiter, so picking 60Hz turned the limiter
on and locked it at 60. Refresh rate now only picks the display mode, and an
explicit rate is no longer steered to a different mode to suit the cap's cadence.

The guest presentation gate is now dropped at boot unconditionally, making the
frame pacer the sole rate authority so a cap can be raised above the panel rate
mid-session. It has to be unconditional: the guest reads its config once at
start, so a session booted without a cap could never raise its ceiling
afterwards. No guest-side cap is emitted alongside it, since a second cap would
beat against the pacer. With no cap set the pacer holds the panel rate rather
than letting the guest free-run.

Drop the vsync-grid alignment. It paced against the rate the display mode was
requested at rather than the rate the panel was actually running, and that
request lands asynchronously and can be refused, so it quantized real vsync
timestamps against a period that did not exist: a 60 cap ran at 80 and a 90 cap
swung between 89 and 100, while caps that bypassed the grid held exact. The
pacer now free-runs off its own accumulator and reads the live display mode for
its no-cap fallback.
